Move all tests in `compile_test/` to a seperate compile-pass test
