--- compile_test/Cargo.toml
@@ -1,7 +1,0 @@
-[package]
-name = "auto_impl_compile_test"
-version = "0.1.0"
-authors = ["Ashley Mannix <ashleymannix@live.com.au>"]
-
-[dependencies]
-auto_impl = { version = "*", path = "../" }
