+++ tests/compile-pass/big_trait_for_all_refs.rs
@@ -1,0 +1,16 @@
+#![feature(use_extern_macros)]
+
+extern crate auto_impl;
+use auto_impl::auto_impl;
+
+
+#[auto_impl(Arc, Box, Rc, &, &mut)]
+trait RefTrait1<'a, T: for<'b> Into<&'b str>> {
+    type Type1;
+    type Type2;
+
+    const FOO: u32;
+
+    fn execute1<'b>(&'a self, arg1: &'b T) -> Result<Self::Type1, String>;
+    fn execute2(&self) -> Self::Type2;
+}
