+++ tests/compile-pass/fn_method_lifetimes.rs
@@ -1,0 +1,15 @@
+#![feature(use_extern_macros)]
+
+extern crate auto_impl;
+use auto_impl::auto_impl;
+
+
+#[auto_impl(Fn)]
+trait FnTrait2<'a, T> {
+    fn execute<'b, 'c>(
+        &'a self,
+        arg1: &'b T,
+        arg2: &'c T,
+        arg3: &'static str,
+    ) -> Result<T, String>;
+}
