+++ tests/compile-pass/ref_for_fn.rs
